core: api-server: websocket: Add authentication to relevant routes
